Refactor domain types
