+++ lib/domain/sight_type.dart
@@ -1,0 +1,9 @@
+/// Перечисление возможных типов мест
+enum SightType {
+  hotel,
+  restaurant,
+  particular_place,
+  park,
+  museum,
+  cafe,
+}
