+++ lib/ui/screens/add_sight/add_sight_logic.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:places/domain/category.dart';
 import 'package:places/domain/sight.dart';
 import 'package:places/mocks.dart';
 import 'package:places/ui/res/strings/strings.dart';
